+++ rsclf/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: Copyright © Idiap Research Institute <contact@idiap.ch>
+#
+# SPDX-License-Identifier: GPL-3.0-only
+
+"""Raw Speech Classification package."""