+++ rsclf/plot.py
@@ -1,21 +1,10 @@
-# coding=utf-8
+# SPDX-FileCopyrightText: Copyright © Idiap Research Institute <contact@idiap.ch>
+#
+# SPDX-FileContributor: Olivier Canévet <olivier.canevet@idiap.ch>
+#
+# SPDX-License-Identifier: GPL-3.0-only
 
-# Copyright (c) 2024 Idiap Research Institute <https://www.idiap.ch>
-# Written by Olivier Canévet <olivier [dot] canevet [at] idiap [dot] ch>
-#
-# This file is part of RawSpeechClassification.
-#
-# RawSpeechClassification is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# RawSpeechClassification is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with RawSpeechClassification. If not, see <http://www.gnu.org/licenses/>.
+"""Provide a command line interface for plotting results of an experiment."""
 
 import argparse
 
@@ -26,6 +15,7 @@ import polars as pl
 
 
 def prepare_figure():
+    """Set up the axes, grid, and labels of the current figure."""
     fig, axs = plt.subplots(2, 2, figsize=(10, 8))
     axs[0, 0].set_ylabel("Training loss")
     axs[0, 0].grid(visible=True, linestyle="--")
@@ -38,11 +28,23 @@ def prepare_figure():
     return fig, axs
 
 
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--output-dir", default=".", help="Output directory")
+def main():
+    """Create plots showing the performance of the classification for each epoch.
+
+    Each given directory must contain a log.dat file and an accuracy.dat file.
+    """
+    parser = argparse.ArgumentParser(prog="rsclf-plot", description=main.__doc__)
+    parser.add_argument(
+        "--output-dir",
+        default=".",
+        help="Output directory for the plots file",
+    )
     parser.add_argument("--extension", default="png", help="Image extension to save")
-    parser.add_argument("dirnames", nargs=argparse.REMAINDER)
+    parser.add_argument(
+        "dirnames",
+        nargs="+",
+        help="Location of the Keras trained model",
+    )
     args = parser.parse_args()
 
     Path(args.output_dir).mkdir(exist_ok=True, parents=True)
@@ -51,9 +53,11 @@ if __name__ == "__main__":
 
     for dirname in args.dirnames:
         log_path = Path(dirname) / "log.dat"
-        assert Path(log_path).is_file(), f"Expect file {log_path}"
+        if not Path(log_path).is_file():
+            raise FileNotFoundError(f"Expect file {log_path}")
         acc_path = Path(dirname) / "accuracy.dat"
-        assert Path(acc_path).is_file(), f"Expect file {acc_path}"
+        if not Path(acc_path).is_file():
+            raise FileNotFoundError(f"Expect file {acc_path}")
 
         logs = pl.read_csv(log_path, has_header=True, separator=" ")
         axs[0, 0].plot(logs["loss"])
@@ -63,3 +67,7 @@ if __name__ == "__main__":
 
     fig.savefig(Path(args.output_dir) / f"plot.{args.extension}", bbox_inches="tight")
     plt.close(fig)
+
+
+if __name__ == "__main__":
+    main()