+++ rsclf/rawdataset.py
@@ -1,25 +1,14 @@
-# coding=utf-8
+# SPDX-FileCopyrightText: Copyright © Idiap Research Institute <contact@idiap.ch>
+#
+# SPDX-FileContributor: S. Pavankumar Dubagunta <pavankumar.dubagunta@idiap.ch>
+# SPDX-FileContributor: Mathew Magimai Doss <mathew@idiap.ch>
+# SPDX-FileContributor: Olivier Bornet <olivier.bornet@idiap.ch>
+# SPDX-FileContributor: Olivier Canévet <olivier.canevet@idiap.ch>
+# SPDX-FileContributor: Yannick Dayer <yannick.dayer@idiap.ch>
+#
+# SPDX-License-Identifier: GPL-3.0-only
 
-# Copyright (c) 2018-2024 Idiap Research Institute <https://www.idiap.ch>
-# Written by S. Pavankumar Dubagunta <pavankumar [dot] dubagunta [at] idiap [dot] ch>
-# and Mathew Magimai Doss <mathew [at] idiap [dot] ch>
-# and Olivier Bornet <olivier [dot] bornet [at] idiap [dot] ch>
-# and Olivier Canévet <olivier [dot] canevet [at] idiap [dot] ch>
-#
-# This file is part of RawSpeechClassification.
-#
-# RawSpeechClassification is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# RawSpeechClassification is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with RawSpeechClassification. If not, see <http://www.gnu.org/licenses/>.
-
+"""Provide an interface to load files from a dataset."""
 
 import pickle
 
@@ -31,10 +20,12 @@ import numpy as np
 
 
 class RawDataset(keras.utils.PyDataset):
+    """Keras dataset to provide samples from a dataset."""
+
     def __init__(self, featDir, batchSize=256, spliceSize=25, mode="train", **kwargs):
         kwargs = {"workers": 1}
         super().__init__(**kwargs)
-        self.featDir = featDir
+        self.featDir = Path(featDir)
         self.batchSize = batchSize
         self.spliceSize = spliceSize
         self.mode = mode
@@ -65,7 +56,7 @@ class RawDataset(keras.utils.PyDataset):
         self.doUpdateSplit = True
 
     def addContextNorm(self, feat):
-        # Add context to get the window size
+        """Add context to get the window size."""
         N = len(feat)
 
         # Repeat feat[0], feat[-1] so that we get the same number of spliced feats
@@ -74,23 +65,26 @@ class RawDataset(keras.utils.PyDataset):
                 np.tile(feat[0], (self.context, 1)),
                 feat,
                 np.tile(feat[-1], (self.context, 1)),
-            ]
+            ],
         )
 
         feat = np.lib.stride_tricks.as_strided(
-            feat, strides=feat.strides, shape=(N, self.inputFeatDim)
+            feat,
+            strides=feat.strides,
+            shape=(N, self.inputFeatDim),
         )
 
         std = feat.std(axis=-1)
         std[std < self.stdFloor] = self.stdFloor
-        feat = ((feat.T - feat.mean(axis=-1)) / std).T
-        return feat
+        return ((feat.T - feat.mean(axis=-1)) / std).T
 
     def __len__(self):
+        """Return the number of steps."""
         return self.numSteps
 
     # Retrieve a mini batch
     def __getitem__(self, idx):
+        """Return a batch at a specific index."""
         self.numDone += 1
         if self.mode == "train":
             while self.batchPointer + self.batchSize >= len(self.x):
@@ -100,8 +94,8 @@ class RawDataset(keras.utils.PyDataset):
 
                 self.splitDataCounter += 1
 
-                featFile = f"{self.featDir}/{self.splitDataCounter}.x.h5"
-                labelFile = f"{self.featDir}/{self.splitDataCounter}.y.h5"
+                featFile = self.featDir / f"{self.splitDataCounter}.x.h5"
+                labelFile = self.featDir / f"{self.splitDataCounter}.y.h5"
 
                 with h5py.File(featFile, "r") as f:
                     featList = [self.addContextNorm(f[i][()]) for i in f]
@@ -130,12 +124,12 @@ class RawDataset(keras.utils.PyDataset):
             self.batchPointer += self.batchSize
             return (xMini, yMini)
 
-        else:  # Test mode
+        else:  # Test mode # noqa: RET505
             while True:
                 if self.doUpdateSplit:
                     self.splitDataCounter += 1
-                    uflFile = f"{self.featDir}/{self.splitDataCounter}.pickle"
-                    self.f = open(uflFile, "rb")
+                    uflFile = self.featDir / f"{self.splitDataCounter}.pickle"
+                    self.f = uflFile.open("rb")
                     self.doUpdateSplit = False
                 try:
                     utt, feat, lab = pickle.load(self.f)