+++ rsclf/train.py
@@ -1,28 +1,17 @@
 #!/usr/bin/env python3
-# coding=utf-8
 
-# Copyright (c) 2018-2024 Idiap Research Institute <https://www.idiap.ch>
-# Written by S. Pavankumar Dubagunta <pavankumar [dot] dubagunta [at] idiap [dot] ch>
-# and Mathew Magimai Doss <mathew [at] idiap [dot] ch>
-# and Olivier Bornet <olivier [dot] bornet [at] idiap [dot] ch>
-# and Olivier Canévet <olivier [dot] canevet [at] idiap [dot] ch>
+# SPDX-FileCopyrightText: Copyright © Idiap Research Institute <contact@idiap.ch>
 #
-# This file is part of RawSpeechClassification.
+# SPDX-FileContributor: S. Pavankumar Dubagunta <pavankumar.dubagunta@idiap.ch>
+# SPDX-FileContributor: Mathew Magimai Doss <mathew@idiap.ch>
+# SPDX-FileContributor: Olivier Bornet <olivier.bornet@idiap.ch>
+# SPDX-FileContributor: Olivier Canévet <olivier.canevet@idiap.ch>
 #
-# RawSpeechClassification is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# RawSpeechClassification is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with RawSpeechClassification. If not, see <http://www.gnu.org/licenses/>.
+# SPDX-License-Identifier: GPL-3.0-only
+
+"""Provide a command line interface for training a model."""
 
 import argparse
-import inspect
 
 from pathlib import Path
 
@@ -31,13 +20,12 @@ import numpy as np
 
 from keras.optimizers import SGD
 
-from model_architecture import model_architecture
-from rawdataset import RawDataset
+from .model_architecture import model_architecture
+from .rawdataset import RawDataset
 
 
 def train(args):
-    """Train the model by running `min-epoch` at a constant learning rate before reducing it."""
-
+    """Train the model by running `min-epoch` at a constant LR before reducing it."""
     tr_dir = args.train_feature_dir
     cv_dir = args.validation_feature_dir
     exp = args.output_dir
@@ -62,20 +50,25 @@ def train(args):
     # Number of times the learning rate has to be scaled.
     learning["lrScaleCount"] = int(
         np.ceil(
-            np.log(learning["minLr"] / learning["rate"]) / np.log(learning["lrScale"])
-        )
+            np.log(learning["minLr"] / learning["rate"]) / np.log(learning["lrScale"]),
+        ),
     )
 
     Path(exp).mkdir(exist_ok=True, parents=True)
     logger = keras.callbacks.CSVLogger(
-        Path(exp) / "log.dat", separator=" ", append=True
+        Path(exp) / "log.dat",
+        separator=" ",
+        append=True,
     )
 
     cvGen = RawDataset(cv_dir, learning["batchSize"], learning["spliceSize"])
     trGen = RawDataset(tr_dir, learning["batchSize"], learning["spliceSize"])
 
     s = SGD(
-        learning_rate=learning["rate"], weight_decay=0, momentum=0.5, nesterov=False
+        learning_rate=learning["rate"],
+        weight_decay=0,
+        momentum=0.5,
+        nesterov=False,
     )
 
     # Initialise model
@@ -130,47 +123,51 @@ def train(args):
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Train and validate the model")
+    """Train and validate the model.
+
+    Set the KERAS_BACKEND environment variable to torch or tensorflow.
+    """
+    parser = argparse.ArgumentParser(prog="rsclf-train", description=main.__doc__)
     # fmt: off
     parser.add_argument(
         "--train-feature-dir", required=True,
-        help="Path to the directory containing the features for training"
+        help="Path to the directory containing the features for training",
     )
     parser.add_argument(
         "--validation-feature-dir", required=True,
-        help="Path to the directory containing the features for validation"
+        help="Path to the directory containing the features for validation",
     )
     parser.add_argument(
-        "--arch", default="seg",
-        help="Model architecture name"
+        "--arch", choices=["seg", "subseg"], default="seg",
+        help="Model architecture name",
     )
     parser.add_argument(
         "--splice-size", type=int, default=25,
-        help="Slice size for feature context"
+        help="Slice size for feature context",
     )
     parser.add_argument(
         "--learning-rate", type=float, default=0.1,
-        help="Initial learning rate"
+        help="Initial learning rate",
     )
     parser.add_argument(
         "--learning-scale", type=float, default=0.5,
-        help="Factor by which to reduce the learning rate"
+        help="Factor by which to reduce the learning rate",
     )
     parser.add_argument(
         "--batch-size", type=int, default=256,
-        help="Batch size"
+        help="Batch size",
     )
     parser.add_argument(
         "--min-epoch", type=int, default=5,
-        help="Minimum epochs to run before reducing learning rate"
+        help="Minimum epochs to run before reducing learning rate",
     )
     parser.add_argument(
         "--output-dir", default="output-results",
-        help="Output directory"
+        help="Output directory",
     )
     parser.add_argument(
         "--verbose", type=int, default=0,
-        help="Keras verbose level for fit and predict"
+        help="Keras verbose level for fit and predict",
     )
     # fmt: on
     args = parser.parse_args()