+++ rsclf/wav2feat.py
@@ -1,54 +1,44 @@
 #!/usr/bin/env python3
-# coding=utf-8
 
-# Copyright (c) 2018-2024 Idiap Research Institute <https://www.idiap.ch>
-# Written by S. Pavankumar Dubagunta <pavankumar [dot] dubagunta [at] idiap [dot] ch>
-# and Mathew Magimai Doss <mathew [at] idiap [dot] ch>
-# and Olivier Canévet <olivier [dot] canevet [at] idiap [dot] ch>
+# SPDX-FileCopyrightText: Copyright © Idiap Research Institute <contact@idiap.ch>
 #
-# This file is part of RawSpeechClassification.
+# SPDX-FileContributor: S. Pavankumar Dubagunta <pavankumar.dubagunta@idiap.ch>
+# SPDX-FileContributor: Mathew Magimai Doss <mathew@idiap.ch>
+# SPDX-FileContributor: Olivier Canévet <olivier.canevet@idiap.ch>
+# SPDX-FileContributor: Yannick Dayer <yannick.dayer@idiap.ch>
 #
-# RawSpeechClassification is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# RawSpeechClassification is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with RawSpeechClassification. If not, see <http://www.gnu.org/licenses/>.
+# SPDX-License-Identifier: GPL-3.0-only
+
+"""Provide a command line interface for extracting features from raw audio files."""
 
 import argparse
-import os
 import pickle
+import wave
 
 from pathlib import Path
 
 import h5py
 import numpy
 import scipy.io.wavfile as wav
-import wave
 
 
 class WAV2featExtractor:
-    """Extractor from wav files to features
+    """Extractor from wav files to features.
 
     Args:
 
-      wavLabListFile (str):
+      wavLabListFile (str): Path to the file containing the list of wav files and labels
       featDir (str): Output directory where to save the features
-      param:
+      param (dict or None): Dictionary containing the parameters for the feature
+        extraction
       mode (str): Mode of the data 'train' or 'test'
-      root (str): Prefix to append to the path
+      root (str): Prefix to append to the path of each file in wavLabListFile
 
     """
 
     def __init__(self, wavLabListFile, featDir=None, param=None, mode="train", root=""):
-
-        self.wavLabListFile = wavLabListFile
-        self.featDir = featDir
+        self.wavLabListFile = Path(wavLabListFile)
+        self.featDir = Path(featDir)
         self.mode = mode
         self.maxSplitDataSize = 100  # Utterances
         self.root = root
@@ -61,26 +51,28 @@ class WAV2featExtractor:
                 "stdFloor": 1e-3,
             }  # Floor on standard deviation
             param["windowLengthSamples"] = int(
-                param["windowLength"] * param["fs"] / 1000.0
+                param["windowLength"] * param["fs"] / 1000.0,
             )
             param["windowShiftSamples"] = int(
-                param["windowShift"] * param["fs"] / 1000.0
+                param["windowShift"] * param["fs"] / 1000.0,
             )
 
         self.param = param
 
-        self.wll = open(self.wavLabListFile)
+        self.wll = self.wavLabListFile.open()
         self.numFeats, self.numUtterances, self.numLabels = self.checkList(
-            self.wavLabListFile
+            self.wavLabListFile,
         )
 
         self.inputFeatDim = self.param["windowLengthSamples"]
         self.outputFeatDim = 1 if self.numLabels == 2 else self.numLabels
 
     def __exit__(self):
+        """Clean up."""
         self.wll.close()
 
     def add_root(self, filename):
+        """Prepend the root path to a filename."""
         if self.root:
             filename = str(Path(self.root) / filename)
 
@@ -90,15 +82,14 @@ class WAV2featExtractor:
         return str(filename)
 
     def extract(self, wavepath):
-        """Feature extraction routine"""
-
+        """Feature extraction routine."""
         # Read data and labels
         fs, data = wav.read(wavepath)
 
         # Append zeros to data if necessary (we add dither later)
         if len(data) < self.param["windowLengthSamples"]:
             data = numpy.concatenate(
-                [data, numpy.zeros(self.param["windowLengthSamples"] - len(data))]
+                [data, numpy.zeros(self.param["windowLengthSamples"] - len(data))],
             )
 
         # Determine the number of frames, each of windowshift length
@@ -119,13 +110,10 @@ class WAV2featExtractor:
         feat += numpy.random.randn(numFeats, self.param["windowLengthSamples"])
 
         # Mean normalise feature matrix
-        feat = (feat.T - feat.mean(axis=-1)).T
+        return (feat.T - feat.mean(axis=-1)).T
 
-        return feat
-
-    def checkList(self, wavLabListFile):
-        """Check files in list and return attributes"""
-
+    def checkList(self, wavLabListFile: Path):
+        """Check files in list and return attributes."""
         print(f"Checking files in {wavLabListFile}")
         labels = set()
         numFeats = 0
@@ -136,16 +124,16 @@ class WAV2featExtractor:
 
             with wave.open(w) as f:
                 # Check number of channels and sampling rate
-                msg = (
-                    f"ERROR: {w} has multiple channels. "
-                    f"Modify the code accordingly and re-run"
-                )
-                assert f.getnchannels() == 1, msg
-                msg = (
-                    f"ERROR: Sampling frequency mismatch with {w}: "
-                    f"expected {self.param['fs']}, got {f.getframerate()}"
-                )
-                assert f.getframerate() == self.param["fs"], msg
+                if f.getnchannels() != 1:
+                    raise ValueError(
+                        f"ERROR: {w} has multiple channels ({f.getnchannels()}). "
+                        "Modify the code accordingly and re-run.",
+                    )
+                if f.getframerate() != self.param["fs"]:
+                    raise ValueError(
+                        f"ERROR: Sampling frequency mismatch with {w}: "
+                        f"expected {self.param['fs']}, got {f.getframerate()}",
+                    )
                 N = f.getnframes()
 
             numFeats += max(
@@ -161,10 +149,9 @@ class WAV2featExtractor:
         return numFeats, numUtterances, numLabels
 
     def prepareFeatDir(self):
-        """Prepare feature directory for training/testing"""
-
+        """Prepare feature directory for training/testing."""
         # Create output directory
-        os.makedirs(self.featDir, exist_ok=False)
+        self.featDir.mkdir(parents=True, exist_ok=False)
         self.numSplit = -(-self.numUtterances // self.maxSplitDataSize)
 
         # Save info
@@ -177,7 +164,7 @@ class WAV2featExtractor:
             "outputFeatDim": self.outputFeatDim,
         }
         print(self.info)
-        infoFile = f"{self.featDir}/info.npy"
+        infoFile = self.featDir / "info.npy"
         numpy.save(infoFile, self.info)
 
         # In case the object is used as iterator before calling this routine
@@ -187,8 +174,7 @@ class WAV2featExtractor:
         self.wll.seek(0)  # For future use
 
     def processUtterance(self, wl):
-        """Process (return) feature and label for one utterance"""
-
+        """Process (return) feature and label for one utterance."""
         if not wl:
             return None, None
         w, label = wl.split()
@@ -197,15 +183,14 @@ class WAV2featExtractor:
         return w, feat, int(label) * numpy.ones(len(feat), dtype=numpy.int32)
 
     def saveNextSplitData(self):
-        """Save a split"""
-
+        """Save a split."""
         lines = [self.wll.readline() for n in range(self.maxSplitDataSize)]
         featLabList = [self.processUtterance(wl) for wl in lines if wl]
 
         if self.mode == "train":
             uttList, featList, labelList = map(list, zip(*featLabList))
-            featFile = f"{self.featDir}/{self.splitDataCounter}.x.h5"
-            labelFile = f"{self.featDir}/{self.splitDataCounter}.y.h5"
+            featFile = self.featDir / f"{self.splitDataCounter}.x.h5"
+            labelFile = self.featDir / f"{self.splitDataCounter}.y.h5"
 
             # Save features
             with h5py.File(featFile, "w") as f:
@@ -218,41 +203,48 @@ class WAV2featExtractor:
                     f.create_dataset(str(i), data=labels, dtype="int32")
 
         else:
-            featFile = f"{self.featDir}/{self.splitDataCounter}.pickle"
-            with open(featFile, "wb") as f:
+            featFile = self.featDir / f"{self.splitDataCounter}.pickle"
+            with featFile.open("wb") as f:
                 for ufl in featLabList:
                     pickle.dump(ufl, f)
 
     def __iter__(self):
-        """Make the object iterable and retrieve one utterance each time"""
+        """Make the object iterable and retrieve one utterance each time."""
         for wl in self.wll:
             yield self.processUtterance(wl)
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Prepare the features")
+    """Extract the features for each raw audio files in a list and save them."""
+    parser = argparse.ArgumentParser(prog="rsclf-wav2feat", description=main.__doc__)
     # fmt: off
     parser.add_argument(
         "--root", default=None,
-        help="Prefix to add in front of path names if provided. (e.g. root=/ssd/dataset/IEMOCAP)"
+        help=(
+            "Prefix to add in front of path names if provided. "
+            "(e.g. root=/ssd/dataset/IEMOCAP)"
+        ),
     )
     parser.add_argument(
         "--wav-list-file", required=True,
-        help="Path to file containing on each row '/path/to/file.wav <label>'"
+        help="Path to file containing on each row '/path/to/file.wav <label>'",
     )
     parser.add_argument(
         "--mode", required=True, choices=["train", "test"],
-        help="Type of data"
+        help="Type of data",
     )
     parser.add_argument(
         "--feature-dir", default="output-features",
-        help="Path where to save the features"
+        help="Path where to save the features",
     )
     # fmt: on
     args = parser.parse_args()
 
     w2f = WAV2featExtractor(
-        args.wav_list_file, featDir=args.feature_dir, mode=args.mode, root=args.root
+        args.wav_list_file,
+        featDir=args.feature_dir,
+        mode=args.mode,
+        root=args.root,
     )
     w2f.prepareFeatDir()