+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<!--
+SPDX-FileCopyrightText: Copyright © Idiap Research Institute <contact@idiap.ch>
+
+SPDX-License-Identifier: GPL-3.0-only
+-->
+
 # Changelog
 
 ## August 2024
@@ -5,6 +11,7 @@
 - Update the code for Keras 3 with PyTorch or Tensorflow backend
 - Add a CSV logger to save the learning process in a text file
 - Add a plot function to draw the learning parameters
+- Setup the project to be published on pypi
 
 ## May 2019